修改消息头部信息 1. tag和flag分别从1字节改为2字节 2. serial从2字节改为4字节 3. 增加4字节req serial字段（用于客户端向服务器应答已收到最大消息序号，用于断线重连补发消息）

+++ corpc/src/corpc_message_server.h
@@ -38,19 +38,19 @@ namespace corpc {
             void setCrypter(std::shared_ptr<Crypter> &crypter) { _crypter = crypter; }
             uint64_t getCreateTime() { return _createTime; }
             
-            void send(int16_t type, bool isRaw, bool needCrypt, uint8_t tag, uint16_t serial, std::shared_ptr<void> msg);
+            void send(int16_t type, bool isRaw, bool needCrypt, uint16_t tag, uint32_t serial, std::shared_ptr<void> msg);
         private:
             MessageServer *_server;
             std::shared_ptr<Crypter> _crypter;
             time_t _createTime;   // 连接创建时间
-            uint16_t _recvSerial; // 接收消息序号（连接建立后从0开始，必须保持连续，包括心跳数据包，不连续则断线）
+            uint32_t _recvSerial; // 接收消息序号（连接建立后从0开始，必须保持连续，包括心跳数据包，不连续则断线）
         public:
             friend class MessageServer;
         };
         
     private:
-        typedef std::function<void(int16_t type, uint8_t tag, std::shared_ptr<google::protobuf::Message>, std::shared_ptr<Connection>)> MessageHandle;
-        typedef std::function<void(int16_t type, uint8_t tag, std::shared_ptr<std::string>, std::shared_ptr<Connection>)> OtherMessageHandle;
+        typedef std::function<void(int16_t type, uint16_t tag, std::shared_ptr<google::protobuf::Message>, std::shared_ptr<Connection>)> MessageHandle;
+        typedef std::function<void(int16_t type, uint16_t tag, std::shared_ptr<std::string>, std::shared_ptr<Connection>)> OtherMessageHandle;
         
         struct RegisterMessageInfo {
             int16_t type;
@@ -63,7 +63,7 @@ namespace corpc {
         struct WorkerTask {
             int16_t type; // 正数类型消息为proto消息，负数类型消息用于系统消息，如：建立连接(-1)、断开连接(-2)
             bool banned; // 屏蔽
-            uint8_t tag; // 客户端向服务器发带tag消息，服务器对这消息应答消息也需带相同的tag（客户端会等待tag消息返回）
+            uint16_t tag; // 客户端向服务器发带tag消息，服务器对这消息应答消息也需带相同的tag（客户端会等待tag消息返回）
             std::shared_ptr<Connection> connection;  // 消息来源的连接，注意：当type为-1时表示新建立连接，当type为-2时表示断开的连接
             std::shared_ptr<void> msg; // 接收到的消息，注意：当type为-1或-2时，msg中无数据
         };
